+++ shotmanager/properties/shots_global_settings.py
@@ -150,7 +150,8 @@ class UAS_ShotManager_ShotsGlobalSettings(PropertyGroup):
 
         for shot in shotList:
             if shot.enabled or props.shotsGlobalSettings.alsoApplyToDisabledShots:
-                if "STORYBOARD" == shot.shotType and shot.isCameraValid():
+                # if "STORYBOARD" == shot.shotType and shot.isCameraValid():
+                if shot.isCameraValid():
                     shot.camera.data.show_passepartout = self.stb_show_passepartout
 
     stb_show_passepartout: BoolProperty(